\name{tm.newsstand-package}
\alias{tm.newsstand-package}
\alias{tm.newsstand}
\docType{package}
\title{
Plugin to the tm package for the Canadian Newsstand databse.
}
\description{
This package serves as a plugin for the tm package to read in news articles from the Canadian NewsStand database and extracting metadata.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tm.newsstand}
\packageIndices{tm.newsstand}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Simon J. Kiss, Wilfrid Laurier University

Maintainer: Simon Kiss <skiss at wlu dot ca>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
